+++ Tests/AccioKitTests/Services/ManifestHandlerServiceTests.swift
@@ -112,7 +112,7 @@ class ManifestCreatorServiceTests: XCTestCase {
         XCTAssertEqual(
             try! String(contentsOf: manifestResource.url),
             """
-                // swift-tools-version:4.2
+                // swift-tools-version:5.0
                 import PackageDescription
 
                 let package = Package(